Moved auth components to sub directory
