+++ front-end/webapp/src/router.js
@@ -1,8 +1,8 @@
 import Vue from 'vue'
 import Router from 'vue-router'
-import RegisterComponent from "./components/Register.vue"
+import RegisterComponent from "./components/auth/Register.vue"
 import HomeComponent from "./components/Home.vue"
-import LoginComponent from "./components/Login.vue"
+import LoginComponent from "./components/auth/Login.vue"
 
 Vue.use(Router);
